Rename client.js to vast_client.js
