+++ src/index.js
@@ -1,5 +1,5 @@
 import { VASTParser } from './parser/parser.js';
-import { VASTClient } from './client.js';
+import { VASTClient } from './vast_client.js';
 import { VASTTracker } from './tracker.js';
 
 export { VASTClient, VASTParser, VASTTracker };
