+++ src/vast_client.js
