+++ test/tracker.js
@@ -1,7 +1,7 @@
 import path from 'path';
 import sinon from 'sinon';
 import should from 'should';
-import { VASTClient } from '../src/client';
+import { VASTClient } from '../src/vast_client';
 import { VASTParser } from '../src/parser/parser';
 import { VASTTracker } from '../src/tracker';
